renamed to AbstracktKernelTest
